+++ src/Test/Inject/TablePresenter.ts
@@ -1,9 +1,14 @@
-import { DummyActionPresenter, DummyDatePresenter } from "@/Test/Mock";
+import {
+  DummyActionPresenter,
+  DummyDatePresenter,
+  DummyStatePresenter,
+} from "@/Test/Mock";
 import { AttributesPresenter } from "@/UI/ServiceInventory/Presenters/AttributesPresenter";
 import { TablePresenter } from "@/UI/ServiceInventory/Presenters/TablePresenter";
 
 export const tablePresenter = new TablePresenter(
   new DummyDatePresenter(),
   new AttributesPresenter(),
-  new DummyActionPresenter()
+  new DummyActionPresenter(),
+  new DummyStatePresenter()
 );