+++ src/Test/Mock/DummyStatePresenter.ts
@@ -1,0 +1,8 @@
+import { ReactElement } from "react";
+import { StatePresenter } from "@/UI/ServiceInventory/Presenters";
+
+export class DummyStatePresenter implements StatePresenter {
+  getForId(): ReactElement | null {
+    return null;
+  }
+}