+++ src/Test/Mock/index.ts
@@ -1,2 +1,3 @@
 export * from "./DummyActionPresenter";
 export * from "./DummyDatePresenter";
+export * from "./DummyStatePresenter";