+++ src/UI/ServiceInventory/Components/index.ts
@@ -3,3 +3,4 @@ export * from "./AttributesList";
 export * from "./DateWithTooltip";
 export * from "./IdWithCopy";
 export * from "./InstanceActions";
+export * from "./InstanceState";