+++ src/UI/ServiceInventory/TableProvider.test.tsx
@@ -1,10 +1,11 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import { TableProvider } from "./TableProvider";
+import { ServiceModel } from "@/Core";
 
 test("InventoryTableContainer is visible", async () => {
   // Arrange
-  render(<TableProvider instances={[]} />);
+  render(<TableProvider instances={[]} serviceEntity={{} as ServiceModel} />);
   // Act
   // Assert
   const container = screen.getByTestId("InventoryTableContainer");