+++ src/UI/ServiceInventory/TableProvider.tsx
@@ -1,9 +1,10 @@
 import React from "react";
 import { KeycloakInstance } from "keycloak-js";
-import { ServiceInstanceModelWithTargetStates } from "@/Core";
+import { ServiceInstanceModelWithTargetStates, ServiceModel } from "@/Core";
 import {
   AttributesPresenter,
   InstanceActionPresenter,
+  InstanceStatePresenter,
   MomentDatePresenter,
   TablePresenter,
 } from "./Presenters";
@@ -13,9 +14,14 @@ import { InstanceSetStateManager } from "./InstanceSetStateManager";
 export interface Props {
   instances: ServiceInstanceModelWithTargetStates[];
   keycloak?: KeycloakInstance;
+  serviceEntity: ServiceModel;
 }
 
-export const TableProvider: React.FC<Props> = ({ instances, keycloak }) => {
+export const TableProvider: React.FC<Props> = ({
+  instances,
+  keycloak,
+  serviceEntity,
+}) => {
   const datePresenter = new MomentDatePresenter();
   const attributesPresenter = new AttributesPresenter();
   const instanceSetStatePresenter = new InstanceSetStateManager(
@@ -27,10 +33,12 @@ export const TableProvider: React.FC<Props> = ({ instances, keycloak }) => {
     keycloak,
     instanceSetStatePresenter
   );
+  const statePresenter = new InstanceStatePresenter(instances, serviceEntity);
   const tablePresenter = new TablePresenter(
     datePresenter,
     attributesPresenter,
-    actionPresenter
+    actionPresenter,
+    statePresenter
   );
   const rows = tablePresenter.createFromInstances(instances);