+++ src/app/ServiceInventory/ServiceInventory.tsx
@@ -139,6 +139,7 @@ const ServiceInventory: React.FunctionComponent<Props> = (props) => {
               <InventoryTable
                 instances={instancesOfCurrentService}
                 keycloak={keycloak}
+                serviceEntity={serviceEntity}
               />
             )}
           </Card>